Support TPM integration on Windows
