+++ aws_signing_helper/tpm_signer_helper_windows.go
@@ -1,0 +1,12 @@
+//go:build windows
+
+package aws_signing_helper
+
+import (
+	tpm2 "github.com/google/go-tpm/tpm2"
+	"io"
+)
+
+func openTPM() (io.ReadWriteCloser, error) {
+	return tpm2.OpenTPM()
+}
