--- aws_signing_helper/tpm_signer_windows.go
@@ -1,11 +1,0 @@
-//go:build windows
-
-package aws_signing_helper
-
-import (
-	"errors"
-)
-
-func GetTPMv2Signer(opts GetTPMv2SignerOpts) (signer Signer, signingAlgorithm string, err error) {
-	return nil, "", errors.New("unable to use tpm v2 signer on windows")
-}
